moved teebox creation to appropiate service
