+++ FairwayFinder/src/FairwayFinder.Core/Features/GolfCourse/Models/FormModels/CourseFormModel.cs
@@ -1,24 +1,24 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace FairwayFinder.Core.Features.CourseManagement.Models.FormModels;
+namespace FairwayFinder.Core.Features.GolfCourse.Models.FormModels;
 
 public class CourseFormModel
 {
-    public long? course_id { get; set; }
+    public long? CourseId { get; set; }
     
     [Display(Name = "Course Name")]
     [Required(ErrorMessage = "Course Name is required.")]
     [StringLength(100, ErrorMessage = "Name cannot be longer than 100 characters.")]
-    public string name { get; set; } = "";
+    public string Name { get; set; } = "";
 
     [Display(Name = "Address")]
     [Required(ErrorMessage = "Address is required.")]
     [StringLength(200, ErrorMessage = "Address cannot be longer than 200 characters.")]
-    public string address { get; set; } = "";
+    public string Address { get; set; } = "";
 
     [Required(ErrorMessage = "Phone number is required.")]
     [Display(Name = "Phone Number")]
     [Phone(ErrorMessage = "Invalid phone number format.")]
     [StringLength(20, ErrorMessage = "Phone number cannot be longer than 20 characters.")]
-    public string phone_number { get; set; } = "";
+    public string PhoneNumber { get; set; } = "";
 }
