+++ FairwayFinder/src/FairwayFinder.Core/Repositories/Interfaces/ITeeboxRepository.cs
@@ -7,4 +7,6 @@ public interface ITeeboxRepository : IBaseRepository
     public Task<Teebox?> GetTeeByIdAsync(long teeboxId);
     public Task<List<Teebox>> GetTeesForCourseAsync(long courseId);
     public Task<Dictionary<string, string>> GetTeesDropdownForCourseAsync(long courseId);
+    public Task<int> InsertNewTeeAsync(Teebox teebox, List<Hole> holes);
+    public Task<bool> UpdateTeeAsync(Teebox tee, List<Hole> holes);
 }
