+++ FairwayFinder/src/FairwayFinder.Core/Repositories/TeeboxRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Dapper;
+using Dapper.Contrib.Extensions;
 using FairwayFinder.Core.Models;
 using FairwayFinder.Core.Repositories.Interfaces;
 using Microsoft.Extensions.Configuration;
@@ -29,5 +30,54 @@ public class TeeboxRepository(IConfiguration configuration) : BasePgRepository(c
         await using var conn = await GetNewOpenConnection();
         var data = await conn.QueryAsync<KeyValuePair<string, string>>(sql, new {courseId});
         return new Dictionary<string, string>(data);
+    }
+    
+    public async Task<int> InsertNewTeeAsync(Teebox teebox, List<Hole> holes)
+    {
+        await using var conn = await GetNewOpenConnection();
+        await using var trans = await conn.BeginTransactionAsync();
+
+        try
+        {
+            var teeId = await conn.InsertAsync(teebox, trans);
+
+            foreach (var item in holes)
+            {
+                item.teebox_id = teeId;
+                item.course_id = teebox.course_id;
+                await conn.InsertAsync(item, trans);
+            }
+            
+            await trans.CommitAsync();
+            return teeId;
+        }
+        catch (Exception ex)
+        {
+            await trans.RollbackAsync();
+            return -1;
+        }
+    }
+    
+    public async Task<bool> UpdateTeeAsync(Teebox tee, List<Hole> holes)
+    {
+        await using var conn = await GetNewOpenConnection();
+        await using var trans = await conn.BeginTransactionAsync();
+        
+        try
+        {
+            await conn.UpdateAsync(tee, trans);
+            foreach (var item in holes)
+            {
+                await conn.UpdateAsync(item, trans);
+            }
+            
+            await trans.CommitAsync();
+            return true;
+        }
+        catch (Exception ex)
+        {
+            await trans.RollbackAsync();
+            return false;
+        }
     }
 }
