+++ FairwayFinder/src/FairwayFinder.Core/Services/Interfaces/ICourseService.cs
@@ -1,4 +1,4 @@
-using FairwayFinder.Core.Features.CourseManagement.Models.FormModels;
+using FairwayFinder.Core.Features.GolfCourse.Models.FormModels;
 using FairwayFinder.Core.Models;
 
 namespace FairwayFinder.Core.Services.Interfaces;
@@ -11,7 +11,4 @@ public interface ICourseService
     Task<List<Course>> CourseSearchByName(string name);
     Task<int> AddCourseAsync(CourseFormModel form);
     Task<bool> UpdateCourseAsync(long courseId, CourseFormModel form);
-    Task<int> AddTeeAsync(long courseId, TeeboxFormModel form);
-    Task<bool> UpdateTeeAsync(long teeboxId, TeeboxFormModel form);
-    
 }
