+++ FairwayFinder/src/FairwayFinder.Core/Services/Interfaces/ITeeboxService.cs
@@ -1,3 +1,4 @@
+using FairwayFinder.Core.Features.GolfCourse.Models.FormModels;
 using FairwayFinder.Core.Models;
 
 namespace FairwayFinder.Core.Services.Interfaces;
@@ -7,4 +8,6 @@ public interface ITeeboxService
     Task<Teebox?> GetTeeByIdAsync(long teeboxId);
     Task<List<Teebox>> GetTeesForCourseAsync(long courseId);
     Task<Dictionary<string, string>> GetTeesDropdownForCourseAsync(long courseId);
+    Task<int> AddTeeAsync(long courseId, TeeboxFormModel form);
+    Task<bool> UpdateTeeAsync(long teeboxId, TeeboxFormModel form);
 }
