+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/AddCourse.cshtml
@@ -1,4 +1,4 @@
-@model FairwayFinder.Core.Features.CourseManagement.Models.FormModels.CourseFormModel
+@model FairwayFinder.Core.Features.GolfCourse.Models.FormModels.CourseFormModel
 
 @{
     ViewData["Title"] = "Add Course";
