+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/AddTee.cshtml
@@ -1,4 +1,4 @@
-﻿@model FairwayFinder.Core.Features.CourseManagement.Models.FormModels.TeeboxFormModel
+﻿@model FairwayFinder.Core.Features.GolfCourse.Models.FormModels.TeeboxFormModel
 
 @{
     ViewData["Title"] = "Add Tee";
