+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model FairwayFinder.Core.Features.CourseManagement.Models.ViewModels.CourseManagementViewModel
+﻿@model FairwayFinder.Core.Features.GolfCourse.Models.ViewModels.CourseManagementViewModel
 
 @{
     ViewData["Title"] = "Golf Courses";
