+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/ViewCourse.cshtml
@@ -1,4 +1,4 @@
-@model FairwayFinder.Core.Features.CourseManagement.Models.ViewModels.CourseViewModel
+@model FairwayFinder.Core.Features.GolfCourse.Models.ViewModels.CourseViewModel
 
 @{
     ViewData["Title"] = $"{Model.Course.course_name}";
