+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Course/ViewTeebox.cshtml
@@ -1,4 +1,4 @@
-@model FairwayFinder.Core.Features.CourseManagement.Models.ViewModels.TeeboxViewModel
+@model FairwayFinder.Core.Features.GolfCourse.Models.ViewModels.TeeboxViewModel
 
 @{
     ViewData["Title"] = $"{Model.Teebox.teebox_name}";
