+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Shared/_CourseForm.cshtml
@@ -1,22 +1,22 @@
-@model FairwayFinder.Core.Features.CourseManagement.Models.FormModels.CourseFormModel
+@model FairwayFinder.Core.Features.GolfCourse.Models.FormModels.CourseFormModel
 
 @await Html.PartialAsync("_MessageHtmxPartial")
 
 <div class="form-floating mb-3">
-    <input asp-for="name" type="text" class="form-control" placeholder="Course Name" autocomplete="off"/>
-    <label asp-for="name"></label>
-    <span asp-validation-for="name" class="text-danger"></span>
+    <input asp-for="Name" type="text" class="form-control" placeholder="Course Name" autocomplete="off"/>
+    <label asp-for="Name"></label>
+    <span asp-validation-for="Name" class="text-danger"></span>
 </div>
 
 <div class="form-floating mb-3">
-    <input asp-for="address" type="text" class="form-control" placeholder="Address" autocomplete="off"/>
-    <label asp-for="address"></label>
-    <span asp-validation-for="address" class="text-danger"></span>
+    <input asp-for="Address" type="text" class="form-control" placeholder="Address" autocomplete="off"/>
+    <label asp-for="Address"></label>
+    <span asp-validation-for="Address" class="text-danger"></span>
 </div>
 
 <div class="form-floating mb-3">
-    <input asp-for="phone_number" type="text" class="form-control" placeholder="Phone Number" autocomplete="off"/>
-    <label asp-for="phone_number"></label>
-    <span asp-validation-for="phone_number" class="text-danger"></span>
+    <input asp-for="PhoneNumber" type="text" class="form-control" placeholder="Phone Number" autocomplete="off"/>
+    <label asp-for="PhoneNumber"></label>
+    <span asp-validation-for="PhoneNumber" class="text-danger"></span>
 </div>
