+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Shared/_HoleForm.cshtml
@@ -1,4 +1,4 @@
-@model FairwayFinder.Core.Features.CourseManagement.Models.FormModels.TeeboxFormModel
+@model FairwayFinder.Core.Features.GolfCourse.Models.FormModels.TeeboxFormModel
 
 <div class="row">
