+++ FairwayFinder/src/FairwayFinder.Web/Areas/Course/Views/Shared/_TeeboxScorecard.cshtml
@@ -1,5 +1,5 @@
 @using FairwayFinder.Core.Models
-@model FairwayFinder.Core.Features.CourseManagement.Models.ViewModels.TeeboxViewModel
+@model FairwayFinder.Core.Features.GolfCourse.Models.ViewModels.TeeboxViewModel
 
 @{
     var front_nine_max = 9;
